+++ bigchaindb/backend/query.py
@@ -361,7 +361,7 @@ def get_validator_set(conn, height):
 
 
 @singledispatch
-def get_validator_set_by_election_id(conn, election_id):
+def get_result_by_election_id(conn, election_id, table):
     """Return a validator set change with the specified election_id
     """
 
@@ -369,13 +369,11 @@ def get_validator_set_by_election_id(conn, election_id):
 
 
 @singledispatch
-def get_asset_tokens_for_public_key(connection, asset_id,
-                                    public_key, operation):
+def get_asset_tokens_for_public_key(connection, asset_id, public_key):
     """Retrieve a list of tokens of type `asset_id` that are owned by the `public_key`.
     Args:
         asset_id (str): Id of the token.
         public_key (str): base58 encoded public key
-        operation: filter transaction based on `operation`
     Returns:
         Iterator of transaction that list given owner in conditions.
     """
