+++ tests/tendermint/test_core.py
@@ -189,7 +189,7 @@ def test_end_block_return_validator_updates(b, init_chain_request):
     resp = app.end_block(RequestEndBlock(height=99))
     assert resp.validator_updates[0] == encode_validator(validator)
 
-    updates = b.get_validator_update()
+    updates = b.is_approved()
     assert updates == []
