+++ tests/tendermint/test_lib.py
@@ -163,11 +163,11 @@ def test_validator_updates(b, validator_pub_key):
                         'update_id': VALIDATOR_UPDATE_ID}
     query.store_validator_update(b.connection, validator_update)
 
-    updates = b.get_validator_update()
+    updates = b.is_approved()
     assert updates == [validator_update['validator']]
 
     b.delete_validator_update()
-    assert b.get_validator_update() == []
+    assert b.is_approved() == []
 
 
 @pytest.mark.bdb
